I just realized that the hash update gaijin did was changing ECS_HASH from using fnv-1 to fnv-1a and other stuff still uses fnv-1

+++ replay/mpi/reflection.cpp
@@ -315,7 +315,7 @@ namespace danet
     idFieldSerializer.readFieldsIndex(bs);
     ReflectionVarMeta *v = varList.head;
     //LOG("Deserializing Vars for %p: ", this);
-    for (uint16_t j = 0; j < numVars; ++j)
+    for (uint16_t j = 0; (uint32_t)j < numVars; ++j)
     {
       v = getVarByPersistentId(idFieldSerializer.getFieldId(j));
       if (v)
